Fix Azure SAS authentication: don't add date header for SAS requests
The date header should only be added for shared key authentication,
not for SAS authentication. SAS tokens include the signature in the
query string and don't require a date header.

This fixes the test failure where SAS DELETE requests were including
a date header that the test server didn't expect.